Refactor app into modular structure with components, pages, and data layers
