+++ src/components/ui/index.js
@@ -1,0 +1,3 @@
+export { default as StatCard } from './StatCard';
+export { default as ReviewCard } from './ReviewCard';
+export { default as PageHeader } from './PageHeader';
