+++ src/data/index.js
@@ -1,0 +1,3 @@
+export * from './mockData';
+export * from './menuConfig.jsx';
+export * from './tableColumns.jsx';
