+++ src/pages/index.js
@@ -1,0 +1,6 @@
+export { default as Dashboard } from './Dashboard';
+export { default as NewReview } from './NewReview';
+export { default as AllReviews } from './AllReviews';
+export { default as Regulations } from './Regulations';
+export { default as Reports } from './Reports';
+export { default as Settings } from './Settings';
